LayoutTests/imported/w3c/web-platform-tests/content-security-policy/media-src/media-src-7_1_2.sub-expected.txt

Video element src attribute must match src list - negative test


PASS Disallowed async video src
PASS Disallowed async video source element
FAIL Test that securitypolicyviolation events are fired assert_equals: expected "http://www2.localhost:8800/media/A4.ogv" but got "http://www2.localhost:8800"
